Extract AttachmentProviderFactory to its own file
